@@ -1,5 +1,5 @@
 module Types
-  class BaseInputObject < GraphQL::Schema::InputObject
+  class BaseField < GraphQL::Schema::Field
     argument_class Types::BaseArgument
   end
 end
