@@ -1,5 +1,9 @@
 module Types
   class QueryType < Types::BaseObject
+    # Add `node(id: ID!) and `nodes(ids: [ID!]!)`
+    include GraphQL::Types::Relay::HasNodeField
+    include GraphQL::Types::Relay::HasNodesField
+
     field :items,
           [Types::ItemType],
           null: false,
@@ -7,12 +11,6 @@ module Types
 
     def items
       Item.preload(:user)
-    end
-
-    field :me, Types::UserType, null: true
-
-    def me
-      context[:current_user]
     end
   end
 end
