@@ -3,6 +3,8 @@ Rails.application.routes.draw do
     mount GraphiQL::Rails::Engine, at: "/graphiql", graphql_path: "/graphql"
   end
   post "/graphql", to: "graphql#execute"
-  root 'library#index'
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
+
+  # Almost every application defines a route for the root path ("/") at the top of this file.
+  # root "articles#index"
 end
