@@ -1,4 +1,4 @@
-class CreateItems < ActiveRecord::Migration[6.0]
+class CreateItems < ActiveRecord::Migration[7.0]
   def change
     create_table :items do |t|
       t.string :title
